requirements.txt
faker
boto3
pandas
psycopg2-binary